fix: imports in user profile actions
