+++ packages/apps/human-app/frontend/src/modules/worker/components/profile/register-address-btn.tsx
@@ -1,6 +1,6 @@
 import { t } from 'i18next';
 import { Button } from '@/shared/components/ui/button';
-import { useRegisterAddress } from '@/modules/worker/hooks/use-register-address/use-register-address';
+import { useRegisterAddress } from '@/modules/worker/hooks/use-register-address';
 import { useRegisterAddressNotifications } from '@/modules/worker/hooks/use-register-address-notifications';
 
 export function RegisterAddressBtn() {
