--- packages/apps/human-app/frontend/src/modules/worker/hooks/use-register-address/index.ts
@@ -1,1 +1,0 @@
-export { useRegisterAddress } from './use-register-address';
